All problems fixed

+++ problem1.cpp
@@ -1,7 +1,7 @@
 /**
  * @file problem1.cpp
  * @author Konner Knoll
- * @brief 
+ * @brief This program calculates how many gallons of paint are needed to paint a fence
  * @version 0.1
  * @date 2025-09-05
  * 
@@ -17,10 +17,10 @@ int main(){
     double height = 6;
     double length = 100;
     double squareFeetPerGallon = 340;
-    double numCoats = 2;
+    double numCoats = 2; // number of coats of paint
 
     // Calculating gallons needed and displaying to console
-    double gallons = numCoats * height * length / squareFeetPerGallon;
+    int gallons = numCoats * height * length / squareFeetPerGallon;
     cout << "Gallons needed: " << gallons << endl;
 
     return 0;
